fix: add ::bigint casts to sqlc.narg() org filters to fix PostgreSQL type inference
Queries using sqlc.narg('organisation_id') without a type cast caused
PostgreSQL error 'could not determine data type of parameter ' because
 IS NULL has no type context in the first clause. Adding ::bigint cast
tells PostgreSQL the parameter type, allowing pgx to pass interface{}
parameters (including nil) correctly.

Affected queries: ListAdminUserAccess, ListPartnerAPIKeys,
ListPartnerWebhookSubscriptions, ListPartnerWebhookEvents,
ListPartnerExportRuns, ListIntegrationStatusChecks,
GetPartnerExportRunForOrganisation, ListAdminAuditEvents,
ListPilotIngestionRuns

+++ services/api/queries/audit.sql
@@ -64,6 +64,6 @@ SELECT
     entity_id,
     metadata
 FROM audit_events
-WHERE sqlc.narg('organisation_id') IS NULL OR organisation_id = sqlc.narg('organisation_id')
+WHERE sqlc.narg('organisation_id')::bigint IS NULL OR organisation_id = sqlc.narg('organisation_id')
 ORDER BY created_at DESC, id DESC
 LIMIT sqlc.arg('limit');
